Updated query and models to adhere hibernate 5

+++ pdf-app/src/main/java/com/bitMiners/pdf/repositories/QuestionTypeRepository.java
@@ -2,6 +2,6 @@ package com.bitMiners.pdf.repositories;
 
 import com.bitMiners.pdf.domain.QuestionType;
 
-public interface QuestionTypeRepository extends CrudRepository<QuestionType, Integer> {
+public interface QuestionTypeRepository extends CrudRepository<QuestionType, Long> {
 
 }
